+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## v0.1
 
 * First version of alan
+* First version of models
+* Populate script
+* Admin page first version